Fixes #11
NPCs now can look at random game objects in the world, at a random
interval, with rotational linear interpolation, all because of a tiny
little script. As a side note, all available game objects are now tagged
for convenience.

And when I say all of them, I mean the ones that needed to be tagged.
The rest are left as is.