Fixes #12
Creates content for all objects missing it. Also implements a hardcoded
solution to prevent skipping dialog in the final scene, where you are
intended to check the computer before checking the phone. The new
hardcoded solution is called "-before", where if the phone hasn't rung
yet, "phonebefore" is used; "phone" is used if it has rung.

Tested and works fine.